fix(jira-scripts): don't react to keystrokes when in Jira's visual editor
Fixing this min. for Firefox, in Chrome these key events are not caught anyway for some reason.

The visual editor is filtered by its attribute "contenteditable". This filtering was implemented in the master branch of the abandoned "jquery.hotkeys" only and in a pretty complicated way -> better use the cleaned-up version in my fork.